+++ src/app/api/models/task-definition.ts
@@ -8,6 +8,8 @@ import {TaskDefinitionService} from '../services/task-definition.service';
 import {Grade, GroupSet, LearningOutcome, Project, TutorialStream, Unit} from './doubtfire-model';
 import {Task} from './doubtfire-model';
 import {TaskPrerequisite} from './task-prerequisite';
+import {DiscussionPrompt} from './discussion-prompt';
+import {OverseerStep} from './overseer/overseer-step';
 
 export type UploadRequirement = {
   key: string;
@@ -57,12 +59,19 @@ export class TaskDefinition extends Entity {
   assessInPortfolioOnly: boolean;
   useResourcesForJplagBaseCode: boolean;
   lockAssessmentsToTutorialStream: boolean;
+  discussionPromptsCount: number;
+  overseerResourceFiles: string[] = [];
 
   public readonly taskPrerequisitesCache: EntityCache<TaskPrerequisite> =
     new EntityCache<TaskPrerequisite>();
 
+  public readonly discussionPromptsCache: EntityCache<DiscussionPrompt> =
+    new EntityCache<DiscussionPrompt>();
+
   public readonly learningOutcomesCache: EntityCache<LearningOutcome> =
     new EntityCache<LearningOutcome>();
+
+  public readonly overseerStepsCache: EntityCache<OverseerStep> = new EntityCache<OverseerStep>();
 
   readonly unit: Unit;