+++ src/app/api/services/overseer-assessment.service.ts
@@ -5,6 +5,7 @@ import {HttpClient} from '@angular/common/http';
 import API_URL from 'src/app/config/constants/apiUrl';
 import {OverseerAssessment} from '../models/overseer/overseer-assessment';
 import {Task} from '../models/doubtfire-model';
+import {OverseerStepResultService} from './overseer-step-result.service';
 
 @Injectable()
 export class OverseerAssessmentService extends EntityService<OverseerAssessment> {
@@ -13,7 +14,10 @@ export class OverseerAssessmentService extends EntityService<OverseerAssessment>
   protected readonly triggerEndpointFormat =
     'projects/:project_id:/task_def_id/:td_id:/overseer_assessment/:id:/trigger';
 
-  constructor(httpClient: HttpClient) {
+  constructor(
+    httpClient: HttpClient,
+    private overseerStepResultService: OverseerStepResultService,
+  ) {
     super(httpClient, API_URL);
 
     this.mapping.addKeys(
@@ -31,6 +35,24 @@ export class OverseerAssessmentService extends EntityService<OverseerAssessment>
         },
       },
       ['timestampString', 'submission_timestamp'],
+      {
+        keys: 'overseerStepResults',
+        toEntityOp: (data: object, key: string, overseerAssesment: OverseerAssessment) => {
+          data[key]?.forEach((overseerStep) => {
+            overseerAssesment.stepResultsCache.getOrCreate(
+              overseerStep['id'],
+              this.overseerStepResultService,
+              overseerStep,
+              {
+                constructorParams: overseerAssesment,
+              },
+            );
+          });
+        },
+      },
+      'overseerStepId',
+      'totalSteps',
+      'passedSteps',
     );
   }
 
@@ -44,7 +66,9 @@ export class OverseerAssessmentService extends EntityService<OverseerAssessment>
       td_id: task.definition.id,
     };
 
-    return this.query(pathIds);
+    return this.query(pathIds, {
+      constructorParams: task,
+    });
   }
 
   public triggerOverseer(assessment: OverseerAssessment): Observable<OverseerAssessment> {